new changes

+++ backend/routes/posts.js
@@ -32,7 +32,8 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -56,13 +57,18 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then(() => {
-        if(req.file) {
-            deletePicture(path);
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if(result.modifiedCount > 0) {
+            if(req.file) {
+                deletePicture(path);
+            }
+            res.status(200).json({ message: 'Update was successful!' });
+        } else {
+            res.status(401).json({ message: 'Not Authorized!' });
         }
-        res.status(200).json({ message: 'Update was successful!' });
     });
 });
 
@@ -100,9 +106,13 @@ router.get('/:id', (req, res, next) => {
 
 router.delete('/:id', checkAuth, (req, res, next) => {
     let path = req.headers.path;
-    Post.deleteOne({ _id: req.params.id }).then(() => {
-        deletePicture(path);
-        res.status(200).json({ message: 'Post deleted!' });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if(result.deletedCount > 0) {
+            deletePicture(path);
+            res.status(200).json({ message: 'Post deleted!' });
+        } else {
+            res.status(401).json({ message: 'Not Authorized!' });
+        }
     });
 });
